add inheritance to concept scheme selectors
